requirements.txt
# Core Framework
Flask==3.1.0
SQLModel==0.0.14
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0

# Configuration & Data
PyYAML==6.0.1
pytz==2024.1

# Authentication & Security
Flask-HTTPAuth==4.8.0
Flask-Session==0.5.0
Werkzeug==3.0.1

# QR Code Generation
qrcode[pil]==7.4.2
Pillow==10.2.0

# API & Integrations
requests==2.31.0
cryptography==41.0.8
user-agents==2.2.0

# System Monitoring
psutil==5.9.8

# Logging
python-json-logger==2.0.7

# Database
alembic==1.13.1

# Validation
email-validator==2.1.0

# Development & Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# Security Testing
safety==3.0.1
bandit==1.7.6

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Optional Features (uncomment as needed)
# Flask-Babel==4.0.0  # For advanced i18n
# Flask-Mail==0.9.1   # For email notifications
# celery==5.3.4       # For async tasks
redis==5.0.1        # For caching
Flask-Caching==2.1.0  # Flask cache extension